refactor: search store
